BFIL POC Updates

+++ app/src/main/java/io/esper/files/activity/MainActivity.kt
@@ -26,6 +26,7 @@ import com.ferfalk.simplesearchview.SimpleSearchView
 import com.ferfalk.simplesearchview.utils.DimensUtils.convertDpToPx
 import hendrawd.storageutil.library.StorageUtil
 import io.esper.files.R
+import io.esper.files.constants.Constants.BfilSyncFolder
 import io.esper.files.constants.Constants.ExternalRootFolder
 import io.esper.files.constants.Constants.InternalCheckerString
 import io.esper.files.constants.Constants.InternalRootFolder
@@ -78,8 +79,8 @@ class MainActivity : AppCompatActivity() {
 
         toolbar = findViewById(R.id.toolbar)
         toolbar!!.title = sharedPrefManaged!!.getString(
-            SHARED_MANAGED_CONFIG_APP_NAME,
-            R.string.app_name.toString()
+                SHARED_MANAGED_CONFIG_APP_NAME,
+                R.string.app_name.toString()
         )
         setSupportActionBar(toolbar)
         if (supportActionBar != null) {
@@ -107,15 +108,15 @@ class MainActivity : AppCompatActivity() {
     private fun initFileListFragment() {
         val listItemsFragment: ListItemsFragment = ListItemsFragment.newInstance(mCurrentPath)
         supportFragmentManager.beginTransaction().replace(R.id.layout_content, listItemsFragment)
-            .commitAllowingStateLoss()
+                .commitAllowingStateLoss()
     }
 
     private fun requestPermission() {
         ActivityCompat.requestPermissions(
-            this, arrayOf(
+                this, arrayOf(
                 Manifest.permission.WRITE_EXTERNAL_STORAGE,
                 Manifest.permission.READ_EXTERNAL_STORAGE
-            ), storagePermission
+        ), storagePermission
         )
     }
 
@@ -176,14 +177,14 @@ class MainActivity : AppCompatActivity() {
             }
             R.id.action_sync -> {
                 if (!InstallUtil.isUnknownSourcesSettingOpen(
-                        this
-                    )
+                                this
+                        )
                 )
                     InstallUtil.checkUnknownSourcesSetting(this)
                 else if (!InstallUtil.isAccessibilitySettingOpen(
-                        AutoInstallService::class.java,
-                        this
-                    )
+                                AutoInstallService::class.java,
+                                this
+                        )
                 )
                     InstallUtil.checkAccessibilitySetting(this, AutoInstallService::class.java)
                 else {
@@ -238,12 +239,12 @@ class MainActivity : AppCompatActivity() {
 
     private fun checkPermission(): Boolean {
         return ContextCompat.checkSelfPermission(
-            this,
-            Manifest.permission.WRITE_EXTERNAL_STORAGE
+                this,
+                Manifest.permission.WRITE_EXTERNAL_STORAGE
         ) == PackageManager.PERMISSION_GRANTED &&
                 ContextCompat.checkSelfPermission(
-                    this,
-                    Manifest.permission.READ_EXTERNAL_STORAGE
+                        this,
+                        Manifest.permission.READ_EXTERNAL_STORAGE
                 ) == PackageManager.PERMISSION_GRANTED
     }
 
@@ -273,9 +274,9 @@ class MainActivity : AppCompatActivity() {
     }
 
     override fun onRequestPermissionsResult(
-        requestCode: Int,
-        permissions: Array<String>,
-        grantResults: IntArray
+            requestCode: Int,
+            permissions: Array<String>,
+            grantResults: IntArray
     ) {
         super.onRequestPermissionsResult(requestCode, permissions, grantResults)
         if (requestCode == storagePermission) {
@@ -293,7 +294,7 @@ class MainActivity : AppCompatActivity() {
 
     private fun startManagedConfigValuesReciever() {
         val myRestrictionsMgr =
-            getSystemService(Context.RESTRICTIONS_SERVICE) as RestrictionsManager
+                getSystemService(Context.RESTRICTIONS_SERVICE) as RestrictionsManager
         val restrictionsFilter = IntentFilter(Intent.ACTION_APPLICATION_RESTRICTIONS_CHANGED)
 
         val restrictionsReceiver = object : BroadcastReceiver() {
@@ -302,27 +303,27 @@ class MainActivity : AppCompatActivity() {
 
                 val newAppName = if (appRestrictions.containsKey(SHARED_MANAGED_CONFIG_APP_NAME))
                     appRestrictions.getString(SHARED_MANAGED_CONFIG_APP_NAME)
-                        .toString() else getString(R.string.app_name)
+                            .toString() else getString(R.string.app_name)
 
                 val showScreenshotsFolder =
-                    if (appRestrictions.containsKey(SHARED_MANAGED_CONFIG_SHOW_SCREENSHOTS))
-                        appRestrictions.getBoolean(SHARED_MANAGED_CONFIG_SHOW_SCREENSHOTS) else false
+                        if (appRestrictions.containsKey(SHARED_MANAGED_CONFIG_SHOW_SCREENSHOTS))
+                            appRestrictions.getBoolean(SHARED_MANAGED_CONFIG_SHOW_SCREENSHOTS) else false
 
                 if (toolbar != null)
                     toolbar!!.title = newAppName
 
                 sharedPrefManaged!!.edit().putString(SHARED_MANAGED_CONFIG_APP_NAME, newAppName)
-                    .apply()
+                        .apply()
                 if (showScreenshotsFolder != (sharedPrefManaged!!.getBoolean(
-                        SHARED_MANAGED_CONFIG_SHOW_SCREENSHOTS,
-                        false
-                    ))
+                                SHARED_MANAGED_CONFIG_SHOW_SCREENSHOTS,
+                                false
+                        ))
                 ) {
                     sharedPrefManaged!!.edit()
-                        .putBoolean(
-                            SHARED_MANAGED_CONFIG_SHOW_SCREENSHOTS,
-                            showScreenshotsFolder
-                        ).apply()
+                            .putBoolean(
+                                    SHARED_MANAGED_CONFIG_SHOW_SCREENSHOTS,
+                                    showScreenshotsFolder
+                            ).apply()
                     refreshItems()
                 }
             }
@@ -333,7 +334,7 @@ class MainActivity : AppCompatActivity() {
     private fun getManagedConfigValues() {
         var restrictionsBundle: Bundle?
         val userManager =
-            getSystemService(Context.USER_SERVICE) as UserManager
+                getSystemService(Context.USER_SERVICE) as UserManager
         restrictionsBundle = userManager.getApplicationRestrictions(packageName)
         if (restrictionsBundle == null) {
             restrictionsBundle = Bundle()
@@ -341,23 +342,23 @@ class MainActivity : AppCompatActivity() {
 
         val newAppName = if (restrictionsBundle.containsKey(SHARED_MANAGED_CONFIG_APP_NAME))
             restrictionsBundle.getString(SHARED_MANAGED_CONFIG_APP_NAME)
-                .toString() else getString(R.string.app_name)
+                    .toString() else getString(R.string.app_name)
 
         val showScreenshotsFolder =
-            if (restrictionsBundle.containsKey(SHARED_MANAGED_CONFIG_SHOW_SCREENSHOTS))
-                restrictionsBundle.getBoolean(SHARED_MANAGED_CONFIG_SHOW_SCREENSHOTS) else false
+                if (restrictionsBundle.containsKey(SHARED_MANAGED_CONFIG_SHOW_SCREENSHOTS))
+                    restrictionsBundle.getBoolean(SHARED_MANAGED_CONFIG_SHOW_SCREENSHOTS) else false
 
         if (toolbar != null)
             toolbar!!.title = newAppName
 
         sharedPrefManaged!!.edit().putString(SHARED_MANAGED_CONFIG_APP_NAME, newAppName).apply()
         if (showScreenshotsFolder != (sharedPrefManaged!!.getBoolean(
-                SHARED_MANAGED_CONFIG_SHOW_SCREENSHOTS,
-                false
-            ))
+                        SHARED_MANAGED_CONFIG_SHOW_SCREENSHOTS,
+                        false
+                ))
         ) {
             sharedPrefManaged!!.edit()
-                .putBoolean(SHARED_MANAGED_CONFIG_SHOW_SCREENSHOTS, showScreenshotsFolder).apply()
+                    .putBoolean(SHARED_MANAGED_CONFIG_SHOW_SCREENSHOTS, showScreenshotsFolder).apply()
             refreshItems()
         }
 
@@ -368,33 +369,22 @@ class MainActivity : AppCompatActivity() {
      * Async Task to download file from URL
      */
     @Suppress(
-        "NULLABILITY_MISMATCH_BASED_ON_JAVA_ANNOTATIONS",
-        "RECEIVER_NULLABILITY_MISMATCH_BASED_ON_JAVA_ANNOTATIONS"
+            "NULLABILITY_MISMATCH_BASED_ON_JAVA_ANNOTATIONS",
+            "RECEIVER_NULLABILITY_MISMATCH_BASED_ON_JAVA_ANNOTATIONS"
     )
-    private class DownloadFile : AsyncTask<String?, String?, String>() {
+    private class DownloadFile : AsyncTask<String?, String?, Boolean>() {
         private var progressDialog: ProgressDialog? = null
         private var fileName: String? = null
-        private var folder: String? = null
         override fun onPreExecute() {
             super.onPreExecute()
             mActivity.window.addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
-            val dir = File(
-                Environment.getExternalStorageDirectory()
-                    .path + File.separator + "esperfiles" + File.separator + "Synced Files"
-            )
-            if (dir.isDirectory) {
-                val children = dir.list()
-                for (i in children.indices) {
-                    File(dir, children[i]).delete()
-                }
-            }
             progressDialog = ProgressDialog(mContext)
             progressDialog!!.setProgressStyle(ProgressDialog.STYLE_HORIZONTAL)
             progressDialog!!.setCancelable(false)
             progressDialog!!.show()
         }
 
-        override fun doInBackground(vararg f_url: String?): String {
+        override fun doInBackground(vararg f_url: String?): Boolean {
             var count: Int
             try {
                 val url = URL(f_url[0])
@@ -403,17 +393,11 @@ class MainActivity : AppCompatActivity() {
                 val lengthOfFile = connection.contentLength
                 val input: InputStream = BufferedInputStream(url.openStream(), 8192)
                 fileName = f_url[0]!!.substring(f_url[0]!!.lastIndexOf('/') + 1, f_url[0]!!.length)
-                val directory = File(
-                    Environment.getExternalStorageDirectory()
-                        .path + File.separator + "esperfiles" + File.separator + "Synced Files" + File.separator
-                )
+                val directory = File(BfilSyncFolder)
                 if (!directory.exists()) {
                     directory.mkdirs()
                 }
-                val output: OutputStream = FileOutputStream(
-                    Environment.getExternalStorageDirectory()
-                        .path + File.separator + "esperfiles" + File.separator + "Synced Files" + File.separator + "temp.zip"
-                )
+                val output: OutputStream = FileOutputStream(BfilSyncFolder + "temp.zip")
                 val data = ByteArray(1024)
                 var total: Long = 0
                 while (input.read(data).also { count = it } != -1) {
@@ -425,30 +409,31 @@ class MainActivity : AppCompatActivity() {
                 output.flush()
                 output.close()
                 input.close()
-                return "Downloaded at: $folder$fileName"
+                return true
             } catch (e: Exception) {
                 Log.e("Error: ", e.message)
             }
-            return "Something went wrong"
+            return false
         }
 
         override fun onProgressUpdate(vararg values: String?) {
             progressDialog!!.progress = values[0]!!.toInt()
         }
 
-        override fun onPostExecute(message: String) {
-            FileUtils.unzipFromSync(
-                mContext,
-                Environment.getExternalStorageDirectory()
-                    .path + File.separator + "esperfiles" + File.separator + "Synced Files" + File.separator + "temp.zip",
-                Environment.getExternalStorageDirectory()
-                    .path + File.separator + "esperfiles" + File.separator + "Synced Files" + File.separator
-            )
-            val file = File(
-                Environment.getExternalStorageDirectory()
-                    .path + File.separator + "esperfiles" + File.separator + "Synced Files" + File.separator + "temp.zip"
-            )
-            file.delete()
+        override fun onPostExecute(syncSuccess: Boolean) {
+            if (syncSuccess) {
+                val dir = File(BfilSyncFolder)
+                if (dir.isDirectory) {
+                    val children = dir.list()
+                    for (i in children.indices) {
+                        if (File(dir, children[i]).name != "temp.zip")
+                            File(dir, children[i]).delete()
+                    }
+                }
+                FileUtils.unzipFromSync(mContext, BfilSyncFolder + "temp.zip", BfilSyncFolder)
+                val file = File(BfilSyncFolder + "temp.zip")
+                file.delete()
+            }
             progressDialog!!.dismiss()
             mActivity.window.clearFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
         }
